requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
einops>=0.7.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-learn>=1.3.0
pillow>=10.0.0
pyyaml>=6.0
tabulate>=0.9.0
tqdm>=4.65.0
tensorboardX>=2.6.0
matplotlib>=3.7.0
